+++ src/PeNet/Structures/Copyright.cs
@@ -30,18 +30,5 @@ namespace PeNet.Structures
         {
             return Encoding.ASCII.GetString(buff, (int) offset, (int) size);
         }
-
-
-        /// <summary>
-        ///     Convert all object properties to strings.
-        /// </summary>
-        /// <returns>String representation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("Copyright\n");
-            sb.Append(ExtensionMethods.PropertiesToString(this, "{0,-10}:\t{1,10:X}\n"));
-
-            return sb.ToString();
-        }
     }
 }
