+++ src/PeNet/Structures/IMAGE_COR20_HEADER.cs
@@ -200,18 +200,5 @@ namespace PeNet.Structures
                 return null;
             }
         }
-
-
-        /// <summary>
-        ///     Convert all object properties to strings.
-        /// </summary>
-        /// <returns>String representation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("IMAGE_COR20_HEADER\n");
-            sb.Append(this.PropertiesToString("{0,-10}:\t{1,10:X}\n"));
-
-            return sb.ToString();
-        }
     }
 }
