+++ src/PeNet/Structures/IMAGE_DATA_DIRECTORY.cs
@@ -35,17 +35,5 @@ namespace PeNet.Structures
             get { return Buff.BytesToUInt32(Offset + 0x4); }
             set { Buff.SetUInt32(Offset + 0x4, value); }
         }
-
-        /// <summary>
-        ///     Convert all object properties to strings.
-        /// </summary>
-        /// <returns>String representation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("IMAGE_DATA_DIRECTORY\n");
-            sb.Append(this.PropertiesToString("{0,-10}:\t{1,10:X}\n"));
-
-            return sb.ToString();
-        }
     }
 }
