+++ src/PeNet/Structures/IMAGE_DEBUG_DIRECTORY.cs
@@ -124,18 +124,5 @@ namespace PeNet.Structures
                 return Encoding.UTF8.GetString(bytes);
             }
         }
-
-
-        /// <summary>
-        ///     Convert all object properties to strings.
-        /// </summary>
-        /// <returns>String representation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("IMAGE_DEBUG_DIRECTORY\n");
-            sb.Append(this.PropertiesToString("{0,-10}:\t{1,10:X}\n"));
-
-            return sb.ToString();
-        }
     }
 }
