+++ src/PeNet/Structures/IMAGE_DOS_HEADER.cs
@@ -230,16 +230,5 @@ namespace PeNet.Structures
             get { return Buff.BytesToUInt32(Offset + 0x3C); }
             set { Buff.SetUInt32(Offset + 0x3C, value); }
         }
-
-        /// <summary>
-        ///     Creates a string representation of all properties.
-        /// </summary>
-        /// <returns>The header properties as a string.</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("IMAGE_DOS_HEADER\n");
-            sb.Append(this.PropertiesToString("{0,-10}:\t{1,10:X}\n"));
-            return sb.ToString();
-        }
     }
 }
