+++ src/PeNet/Structures/METADATASTREAMHDR.cs
@@ -76,17 +76,5 @@ namespace PeNet.Structures
         {
             return Buff.GetCString(nameOffset);
         }
-
-        /// <summary>
-        ///     Convert all object properties to strings.
-        /// </summary>
-        /// <returns>String representation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("METADATASTREAMHDR\n");
-            sb.Append(this.PropertiesToString("{0,-10}:\t{1,10:X}\n"));
-
-            return sb.ToString();
-        }
     }
 }
