+++ src/PeNet/Structures/METADATASTREAM_US.cs
@@ -58,15 +58,6 @@ namespace PeNet.Structures
             return UserStringsAndIndices.FirstOrDefault(x => x.Item2 == index)?.Item1;
         }
 
-        public override string ToString()
-        {
-            var sb = new StringBuilder("METADATASTREAM_US\n");
-            sb.Append(this.PropertiesToString("{0,-15}:\t{1,10:X}\n"));
-            foreach (var strings in UserStringsAndIndices)
-                sb.Append($"{strings.Item2}\t{strings.Item1}");
-            return sb.ToString();
-        }
-
         private List<Tuple<string, uint>> ParseUserStringsAndIndices()
         {
             var stringsAndIncides = new List<Tuple<string, uint>>();
