+++ src/PeNet/Structures/UNWIND_CODE.cs
@@ -48,17 +48,5 @@ namespace PeNet.Structures
             get { return Buff.BytesToUInt16(Offset + 0x2); }
             set { Buff.SetUInt16(Offset + 0x2, value); }
         }
-
-        /// <summary>
-        ///     Creates a string representation of the objects
-        ///     properties.
-        /// </summary>
-        /// <returns>UNWIND_CODE properties as a string.</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder("UNWIND_CODE\n");
-            sb.Append(this.PropertiesToString("{0,-20}:\t{1,10:X}\n"));
-            return sb.ToString();
-        }
     }
 }
